refactor(library): remove PrivateGameCard wrapper, inline in PrivateGamesClient
PrivateGameCard was a thin wrapper around MeepleCard with no reuse outside
PrivateGamesClient. Inline the component logic directly into the consumer
and delete the standalone file and its unit test (coverage preserved via the
PrivateGamesClient integration test, updated to mock MeepleCard instead).
